targetfs: add all dunfell supported devices
Initially requested by a customer, this is the latest LTS driver
compiled for all devices supported for the dunfell release against the
dunfell sysroots.

It is designed to work with the the pvr gallium frontend. As such a
powervr/* branch of mesa should be selected. Earlier branches using the
pvr dri interface will be unable to work with Wayland EGL.

Signed-off-by: Randolph Sapp <rs@ti.com>